Fix MD federal_eitc_without_age_minimum test pollution
The variable mutated the shared tax_benefit_system parameter tree to
override the federal EITC minimum age, relying on partial cache
invalidation to revert. With the top-level instant cache left stale,
results were order-dependent and wrong when another test populated the
cache first (see #8108).

Rename to md_federal_eitc_without_age_minimum and recompose from
eitc_phased_in, eitc_maximum, and eitc_reduction, preserving the
non-demographic § 32 gates (investment income, SSN). No shared state
is mutated.

Closes #8108
