+++ changelog.d/fix-md-eitc-age-override.fixed.md
@@ -1,0 +1,1 @@
+Maryland unmarried childless EITC: rewrite `federal_eitc_without_age_minimum` (renamed to `md_federal_eitc_without_age_minimum`) to compose the federal EITC without mutating the shared tax-benefit system, eliminating an order-dependent test pollution bug.
