+++ policyengine_us/tests/policy/baseline/gov/states/md/tax/income/credits/eitc/md_federal_eitc_without_age_minimum.yaml
@@ -7,4 +7,4 @@
     state_code: MD
   output:
     eitc: 0
-    federal_eitc_without_age_minimum: 382.50
+    md_federal_eitc_without_age_minimum: 382.50
