+++ policyengine_us/tests/policy/baseline/gov/states/md/tax/income/credits/eitc/md_unmarried_childless_non_refundable_eitc.yaml
@@ -2,7 +2,7 @@
   period: 2021
   input:
     md_qualifies_for_unmarried_childless_eitc: true
-    federal_eitc_without_age_minimum: 500
+    md_federal_eitc_without_age_minimum: 500
     state_code: MD
   output:
     md_unmarried_childless_non_refundable_eitc: 500 # 500*100%
@@ -11,7 +11,7 @@
   period: 2021
   input:
     md_qualifies_for_unmarried_childless_eitc: true
-    federal_eitc_without_age_minimum: 600
+    md_federal_eitc_without_age_minimum: 600
     state_code: MD
   output:
     md_unmarried_childless_non_refundable_eitc: 530 # min(530, 600*100%)
@@ -20,7 +20,7 @@
   period: 2021
   input:
     md_qualifies_for_unmarried_childless_eitc: false
-    federal_eitc_without_age_minimum: 500
+    md_federal_eitc_without_age_minimum: 500
     state_code: MD
   output:
     md_unmarried_childless_non_refundable_eitc: 0
